% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fast_install}
\alias{fast_install}
\title{Install R package dependencies from DESCRIPTION file}
\usage{
fast_install(filename, url = TRUE, verbose = TRUE)
}
\arguments{
\item{filename}{path to DESCRIPTION file.}

\item{url}{specified if filename is an url: must be raw content, if from github, must be specified as raw.githubusercontent.com/PATH}

\item{verbose}{optional text output of installation status.}
}
\description{
Install R package dependencies from DESCRIPTION file
}
