% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fast_install_single_package}
\alias{fast_install_single_package}
\title{Install a single package}
\usage{
fast_install_single_package(pkg, verbose = FALSE, ind = 57, ...)
}
\arguments{
\item{pkg}{package name to be installed.}

\item{verbose}{optional text input to display installation status.}

\item{ind}{index for CRAN mirror selection}

\item{...}{optional parameters for install.packages()}
}
\description{
Install a single package
}
