% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_deps}
\alias{get_deps}
\title{Get list of package dependencies from DESCRIPTION file}
\usage{
get_deps(filename, url = TRUE)
}
\arguments{
\item{filename}{path to DESCRIPTION file.}

\item{url}{specified if filename is an url.}
}
\description{
Get list of package dependencies from DESCRIPTION file
}
